requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
dnspython==2.7.0
docker==7.1.0
faiss-cpu==1.10.0
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
ftfy==6.3.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.20.1
open_clip_torch==2.30.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pymongo==4.11.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
sympy==1.13.1
tensorrt==10.8.0.43
tensorrt_cu12==10.8.0.43
tensorrt_cu12_bindings==10.8.0.43
tensorrt_cu12_libs==10.8.0.43
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
